apps/functions/requirements.txt
azure-functions
httpx>=0.27
python-dotenv>=1.0
